test: DCI 패턴을 적용한 MemberServiceTest 적용
